revno: 11164 message:   Using double column type for mysql
committer: Lars Helge Øverland <larshelge@gmail.com>
branch nick: dhis2
timestamp: Thu 2013-06-06 18:59:23 +0200

+++ dhis-2/dhis-support/dhis-support-jdbc/src/main/java/org/hisp/dhis/jdbc/statementbuilder/MySQLStatementBuilder.java
@@ -38,7 +38,7 @@ public class MySQLStatementBuilder
     @Override
     public String getDoubleColumnType()
     {
-        return "decimal";
+        return "double";
     }
 
     @Override
